requirements.txt
playwright==1.41.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-html==4.1.1
python-dotenv==1.0.0
requests==2.31.0
Faker==0.7.4